+++ src/main/java/chess/domain/board/State.java
@@ -1,4 +1,4 @@
-package chess.board;
+package chess.domain.board;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
